+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperPrBalMatRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperPrBalMat;
 
-public interface TaPaperPrBalMatRepository extends CommonJpaCrudRepository<TaPaperPrBalMat, BigDecimal> {
+public interface TaPaperPrBalMatRepository extends CommonJpaCrudRepository<TaPaperPrBalMat, Long> {
 
 }
