+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperPrInMatRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperPrInMat;
 
-public interface TaPaperPrInMatRepository extends CommonJpaCrudRepository<TaPaperPrInMat, BigDecimal> {
+public interface TaPaperPrInMatRepository extends CommonJpaCrudRepository<TaPaperPrInMat, Long> {
 
 }
