+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperPrRedTaxRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperPrRedTax;
 
-public interface TaPaperPrRedTaxRepository extends CommonJpaCrudRepository<TaPaperPrRedTax, BigDecimal> {
+public interface TaPaperPrRedTaxRepository extends CommonJpaCrudRepository<TaPaperPrRedTax, Long> {
 
 }
