+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaPaperPrUnpriRedTaxRepository.java
@@ -1,10 +1,9 @@
 
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaPaperPrUnpriRedTax;
 
-public interface TaPaperPrUnpriRedTaxRepository extends CommonJpaCrudRepository<TaPaperPrUnpriRedTax, BigDecimal> {
+public interface TaPaperPrUnpriRedTaxRepository extends CommonJpaCrudRepository<TaPaperPrUnpriRedTax, Long> {
 
 }
